Refactor filter controls into own file and add WP filter
